EBS Removal feature implemented
Extended Black Screens (EBS; previously "blackscreens") are a way to load into the game and cause a black screen that lasts for roughly 5-6s. During this blackscreen (after about a 0.5s delay), the game is fully simulating in the background, so things like dialogue timers, spawn timers, etc, are all running as if the game isn't loading. With the way our autosplitter detects loading, EBS time was counted as well. This feature injects the time back into Game Time, but adjusted a bit to account for hardware differences and the fact that a 0.5s blackscreen always happens.