Removed support for Bethesda client
Bethesda sunsetted their launcher on May 11, 2022. You are no longer able to launch games from the client, and need to transfer your library to a Steam account if you wish to keep playing.
If, for some reason, the ASL with Bethesda support is needed, it has been added to the archive folder in this repository.